Move BitcoinTestUtils to co.rsk.peg.bitcoin package
